# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Superconducting Circuits API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_tasks_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "completed",
                    "failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "lane",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "simulation",
                    "characterization"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lane"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "workspace",
                "owned"
              ],
              "type": "string",
              "default": "workspace",
              "title": "Scope"
            }
          },
          {
            "name": "dataset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dataset Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskSummaryResponse"
                  },
                  "title": "Response List Tasks Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Task",
        "operationId": "submit_task_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/events": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Task Events",
        "operationId": "list_task_events_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Order"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "task_submitted",
                    "task_running",
                    "task_completed",
                    "task_failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskEventResponse"
                  },
                  "title": "Response List Task Events Tasks  Task Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetadataRecordRefResponse": {
        "properties": {
          "backend": {
            "type": "string",
            "const": "sqlite_metadata",
            "title": "Backend"
          },
          "record_type": {
            "type": "string",
            "enum": [
              "dataset",
              "trace_batch",
              "analysis_run",
              "result_handle"
            ],
            "title": "Record Type"
          },
          "record_id": {
            "type": "string",
            "title": "Record Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          }
        },
        "type": "object",
        "required": [
          "backend",
          "record_type",
          "record_id",
          "version",
          "schema_version"
        ],
        "title": "MetadataRecordRefResponse"
      },
...
      "TaskDetailResponse": {
        "properties": {
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "simulation",
              "post_processing",
              "characterization"
            ],
            "title": "Kind"
          },
          "lane": {
            "type": "string",
            "enum": [
              "simulation",
              "characterization"
            ],
            "title": "Lane"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "run",
              "smoke"
            ],
            "title": "Execution Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "title": "Submitted At"
          },
          "owner_user_id": {
            "type": "string",
            "title": "Owner User Id"
          },
          "owner_display_name": {
            "type": "string",
            "title": "Owner Display Name"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "workspace_slug": {
            "type": "string",
            "title": "Workspace Slug"
          },
          "visibility_scope": {
            "type": "string",
            "enum": [
              "workspace",
              "owned"
            ],
            "title": "Visibility Scope"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "definition_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition Id"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "queue_backend": {
            "type": "string",
            "const": "in_memory_scaffold",
            "title": "Queue Backend"
          },
          "worker_task_name": {
            "type": "string",
            "enum": [
              "simulation_run_task",
              "simulation_smoke_task",
              "simulation_failure_task",
              "simulation_crash_task",
              "post_processing_run_task",
              "post_processing_smoke_task",
              "characterization_run_task",
              "characterization_smoke_task",
              "characterization_failure_task",
              "characterization_crash_task"
            ],
            "title": "Worker Task Name"
          },
          "request_ready": {
            "type": "boolean",
            "title": "Request Ready"
          },
          "submitted_from_active_dataset": {
            "type": "boolean",
            "title": "Submitted From Active Dataset"
          },
          "dispatch": {
            "$ref": "#/components/schemas/TaskDispatchResponse"
          },
          "progress": {
            "$ref": "#/components/schemas/TaskProgressResponse"
          },
          "result_refs": {
            "$ref": "#/components/schemas/TaskResultRefsResponse"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TaskEventResponse"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "kind",
          "lane",
          "execution_mode",
          "status",
          "submitted_at",
          "owner_user_id",
          "owner_display_name",
          "workspace_id",
          "workspace_slug",
          "visibility_scope",
          "dataset_id",
          "definition_id",
          "summary",
          "queue_backend",
          "worker_task_name",
          "request_ready",
          "submitted_from_active_dataset",
          "dispatch",
          "progress",
          "result_refs",
          "events"
        ],
        "title": "TaskDetailResponse"
      },
      "TaskDispatchResponse": {
        "properties": {
          "dispatch_key": {
            "type": "string",
            "title": "Dispatch Key"
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "submission_source": {
            "type": "string",
            "enum": [
              "active_dataset",
              "explicit_dataset",
              "definition_only"
            ],
            "title": "Submission Source"
          },
          "accepted_at": {
            "type": "string",
            "title": "Accepted At"
          },
          "last_updated_at": {
            "type": "string",
            "title": "Last Updated At"
          }
        },
        "type": "object",
        "required": [
          "dispatch_key",
          "status",
          "submission_source",
          "accepted_at",
          "last_updated_at"
        ],
        "title": "TaskDispatchResponse"
      },
      "TaskEventResponse": {
        "properties": {
          "event_key": {
            "type": "string",
            "title": "Event Key"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "task_submitted",
              "task_running",
              "task_completed",
              "task_failed"
            ],
            "title": "Event Type"
          },
          "level": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "error"
            ],
            "title": "Level"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "event_key",
          "event_type",
          "level",
          "occurred_at",
          "message",
          "metadata"
        ],
        "title": "TaskEventResponse"
      },
      "TaskMutationResponse": {
        "properties": {
          "operation": {
            "type": "string",
            "const": "submitted",
            "title": "Operation"
          },
          "task": {
            "$ref": "#/components/schemas/TaskDetailResponse"
          }
        },
        "type": "object",
        "required": [
          "operation",
          "task"
        ],
        "title": "TaskMutationResponse"
      },
      "TaskProgressResponse": {
        "properties": {
          "phase": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Phase"
          },
          "percent_complete": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percent Complete"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "phase",
          "percent_complete",
          "summary",
          "updated_at"
        ],
        "title": "TaskProgressResponse"
      },
      "TaskResultRefsResponse": {
        "properties": {
          "trace_batch_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Batch Id"
          },
          "analysis_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Run Id"
          },
          "metadata_records": {
            "items": {
              "$ref": "#/components/schemas/MetadataRecordRefResponse"
            },
            "type": "array",
            "title": "Metadata Records"
          },
          "trace_payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TracePayloadRefResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "result_handles": {
            "items": {
              "$ref": "#/components/schemas/ResultHandleRefResponse"
            },
            "type": "array",
            "title": "Result Handles"
          }
        },
        "type": "object",
        "required": [
          "trace_batch_id",
          "analysis_run_id",
          "metadata_records",
          "trace_payload",
          "result_handles"
        ],
        "title": "TaskResultRefsResponse"
      },
      "TaskSubmissionRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "simulation",
              "post_processing",
              "characterization"
            ],
            "title": "Kind"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "definition_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition Id"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "TaskSubmissionRequest"
      },
      "TaskSummaryResponse": {
        "properties": {
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "simulation",
              "post_processing",
              "characterization"
            ],
            "title": "Kind"
          },
          "lane": {
            "type": "string",
            "enum": [
              "simulation",
              "characterization"
            ],
            "title": "Lane"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "run",
              "smoke"
            ],
            "title": "Execution Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "title": "Submitted At"
          },
          "owner_user_id": {
            "type": "string",
            "title": "Owner User Id"
          },
          "owner_display_name": {
            "type": "string",
            "title": "Owner Display Name"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "workspace_slug": {
            "type": "string",
            "title": "Workspace Slug"
          },
          "visibility_scope": {
            "type": "string",
            "enum": [
              "workspace",
              "owned"
            ],
            "title": "Visibility Scope"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "definition_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition Id"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "kind",
          "lane",
          "execution_mode",
          "status",
          "submitted_at",
          "owner_user_id",
          "owner_display_name",
          "workspace_id",
          "workspace_slug",
          "visibility_scope",
          "dataset_id",
          "definition_id",
          "summary"
        ],
        "title": "TaskSummaryResponse"
      },
      "TracePayloadRefResponse": {
        "properties": {
          "contract_version": {
            "type": "string",
            "title": "Contract Version"
          },
          "backend": {
            "type": "string",
            "enum": [
              "local_zarr",
              "s3_zarr"
            ],
            "title": "Backend"
          },
          "payload_role": {
            "type": "string",
            "enum": [
              "dataset_primary",
              "task_output",
              "analysis_projection"
            ],
            "title": "Payload Role"
          },
          "store_key": {
            "type": "string",
            "title": "Store Key"
          },
          "store_uri": {
            "type": "string",
            "title": "Store Uri"
          },
          "group_path": {
            "type": "string",
            "title": "Group Path"
          },
          "array_path": {
            "type": "string",
            "title": "Array Path"
          },
          "dtype": {
            "type": "string",
            "title": "Dtype"
          },
          "shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Shape"
          },
          "chunk_shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Chunk Shape"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          }
        },
        "type": "object",
        "required": [
          "contract_version",
          "backend",
          "payload_role",
          "store_key",
          "store_uri",
          "group_path",
          "array_path",
          "dtype",
          "shape",
          "chunk_shape",
          "schema_version"
        ],
        "title": "TracePayloadRefResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}